+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.1]
 ### Added
 - [issues/50](https://github.com/podaac/l2ss-py/issues/50): Spatial bounds are computed correctly for grouped empty subset operations
+- Added `timeout` option to `cmr-umm-updater`
 ### Changed 
-- Upgraded `cmr-umm-updater` to 0.2.0
+- Upgraded `cmr-umm-updater` to 0.2.1
 ### Deprecated 
 ### Removed
 ### Fixed